Moved some custom Linq to Specifications. Added slnx file. Completed account opening date.
