+++ Asm.MooBank.sln
@@ -200,7 +200,6 @@ Global
 		{52A21417-36F4-416C-A3D5-86D2DFA5F63C}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{52A21417-36F4-416C-A3D5-86D2DFA5F63C}.Release|Any CPU.Build.0 = Release|Any CPU
 		{E7192FDB-1422-4738-8262-7526F4EFF780}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E7192FDB-1422-4738-8262-7526F4EFF780}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{E7192FDB-1422-4738-8262-7526F4EFF780}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{E7192FDB-1422-4738-8262-7526F4EFF780}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
@@ -219,14 +218,14 @@ Global
 		{C5167967-0DA8-4133-A89B-1C246C8536C5} = {38BEA81E-F80F-44DE-A930-9BD5D3A62FFF}
 		{2E2085E6-C45F-4EAB-9F39-8DC83CE3B9B2} = {38BEA81E-F80F-44DE-A930-9BD5D3A62FFF}
 		{609F2607-9FFD-4DF1-AB84-5E745FD6395E} = {EAC5F111-34BB-4633-A274-257074F46E7F}
-		{5EED26F8-2233-49E7-A302-82EF46DACE1A} = {2754ACBC-340B-4D5C-AEE8-1AC1AD5CF9E0}
-		{E9177C4B-593F-4EDE-B782-4CB7BC7C2D36} = {2754ACBC-340B-4D5C-AEE8-1AC1AD5CF9E0}
-		{34A02BFD-A9AC-47EE-A1DE-A5B00C64CCAA} = {38BEA81E-F80F-44DE-A930-9BD5D3A62FFF}
-		{2238CEE0-0A21-4C0A-8A18-2DD08020AF99} = {2754ACBC-340B-4D5C-AEE8-1AC1AD5CF9E0}
-		{544C51D0-794B-426C-B7D4-E1DA45CBE916} = {38BEA81E-F80F-44DE-A930-9BD5D3A62FFF}
-		{511708F0-5282-4B40-9848-792C0E139FF8} = {2754ACBC-340B-4D5C-AEE8-1AC1AD5CF9E0}
-		{60087231-E21C-4E94-884D-285BE13489A7} = {54709F25-BC08-4B76-9DEE-88687AD76533}
-		{E0427C53-3EC8-4D3D-B9EB-8638FB3889A3} = {38BEA81E-F80F-44DE-A930-9BD5D3A62FFF}
+		{5EED26F8-2233-49E7-A302-82EF46DACE1A} = {EAC5F111-34BB-4633-A274-257074F46E7F}
+		{E9177C4B-593F-4EDE-B782-4CB7BC7C2D36} = {38BEA81E-F80F-44DE-A930-9BD5D3A62FFF}
+		{34A02BFD-A9AC-47EE-A1DE-A5B00C64CCAA} = {2754ACBC-340B-4D5C-AEE8-1AC1AD5CF9E0}
+		{2238CEE0-0A21-4C0A-8A18-2DD08020AF99} = {38BEA81E-F80F-44DE-A930-9BD5D3A62FFF}
+		{544C51D0-794B-426C-B7D4-E1DA45CBE916} = {2754ACBC-340B-4D5C-AEE8-1AC1AD5CF9E0}
+		{511708F0-5282-4B40-9848-792C0E139FF8} = {54709F25-BC08-4B76-9DEE-88687AD76533}
+		{60087231-E21C-4E94-884D-285BE13489A7} = {38BEA81E-F80F-44DE-A930-9BD5D3A62FFF}
+		{E0427C53-3EC8-4D3D-B9EB-8638FB3889A3} = {2754ACBC-340B-4D5C-AEE8-1AC1AD5CF9E0}
 		{E7192FDB-1422-4738-8262-7526F4EFF780} = {38BEA81E-F80F-44DE-A930-9BD5D3A62FFF}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
