+++ Directory.Packages.props
@@ -34,13 +34,13 @@
     <PackageVersion Include="System.Text.Json" Version="9.0.0" />
   </ItemGroup>
   <ItemGroup>
-    <PackageVersion Include="Asm" Version="3.4.319" />
-    <PackageVersion Include="Asm.AspNetCore" Version="3.4.319" />
+    <PackageVersion Include="Asm" Version="3.4.321" />
+    <PackageVersion Include="Asm.AspNetCore" Version="3.4.321" />
     <PackageVersion Include="Asm.AspNetCore.Api" Version="3.4.318" />
-    <PackageVersion Include="Asm.Cqrs.AspNetCore" Version="3.4.319" />
-    <PackageVersion Include="Asm.Domain" Version="3.4.319" />
-    <PackageVersion Include="Asm.Domain.Infrastructure" Version="3.4.319" />
-    <PackageVersion Include="Asm.Testing" Version="3.4.319" />
+    <PackageVersion Include="Asm.Cqrs.AspNetCore" Version="3.4.321" />
+    <PackageVersion Include="Asm.Domain" Version="3.4.321" />
+    <PackageVersion Include="Asm.Domain.Infrastructure" Version="3.4.321" />
+    <PackageVersion Include="Asm.Testing" Version="3.4.321" />
   </ItemGroup>
   <ItemGroup>
     <PackageVersion Include="Bogus" Version="35.6.1" />
