+++ src/Asm.MooBank.Domain/Entities/Account/Events/AccountAddedEvent.cs
@@ -1,3 +1,3 @@
 ﻿namespace Asm.MooBank.Domain.Entities.Account.Events;
-internal record AccountAddedEvent(InstitutionAccount Account, decimal OpeningBalance, DateTime openingDate) : IDomainEvent;
+internal record AccountAddedEvent(InstitutionAccount Account, decimal OpeningBalance, DateTime OpeningDate) : IDomainEvent;
