+++ src/Asm.MooBank.Models/Asm.MooBank.Models.csproj
@@ -1,3 +1,5 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
-
+  <ItemGroup>
+    <PackageReference Include="Asm" />
+  </ItemGroup>
 </Project>
